# spirv-val: Optimize struct field decoration lookup