@ShabbyX 2022-06-01
I see glslc tests have a few failures where `OpLine` is now emitted unexpectedly. Is that a flake? Doesn't seem related to this change.