@dneto0 2022-06-03
> I see glslc tests have a few failures where OpLine is now emitted unexpectedly. Is that a flake? Doesn't seem related to this change.

It's not a flake.  

The smoke tests deliberately use top of tree Glslang and Shaderc.  Glslang now has more output when compiling with debug info turned on, and Shaderc did not catch up yet.  I updated Shaderc yesterday. 
https://github.com/google/shaderc/pull/1249